refactor(Balance): Minor Kotlin tidying
